Large rework of error handling and fix for serious time complexity issue This is the first in what is a probably going to be a series of changes aimed at cleaning up and improving the codebase.
This commit removes exceptions and addresses a rather large bug.

The bug in question relates to `lex.position()` recalculating the line
and column number for _every_ AST node insertion. This means the string would
be iterated from beginning to current position for every insertion of a node,
this would lead to pretty abysmal performance in a lot of cases.

The new system completely seperates position info from AST nodes and also
fixes the bug of needless recalculating of line/column info. Positions are
now inserted into a map with the node id as the key. This should keep the AST
a bit more local and faster for traversal too. Position info is only calculated
when an error is printed now.

Other changes include a rework of environment and lexer interfaces which leads
the code to being a bit more readable.