Better errors and improved logging
Made some serious improvements to errors in the form of source snippets
and colours.

Logging is now highly improved also.

Minor improvements to parsing code so that error positions should line up
with what the error is referring to.

Added handling for EOF in error messages.

Position structs now store a View rather than just a pointer.

Fixed an obscure bug where the first call to advance in the lexer constructor
could error and result in the lookahead token's position info being passed
to the error handling code. This would result in a segfault because the
lookahead token is zero-initialised with a nullptr. Fix is to initialise the
lookahead token to the base ptr with a length of one.