Stack operations
Added a global stack that can be pushed to and popped from.

Push is a statement that consumes an expression, evaluates
it and then pushes the string to the stack.

Pop is an expression that evaluates to the top element of
the stack and then pops the top value of the stack.