Code reorganisation & in-progress changes to pop expressions
Pop expressions are now going to call a function by name specified
by the user. The evaluator will check that the function has an overload
for with and without passing the popped value. If the stack is empty,
pop will return "" and otherwise it will call the user supplied function
and return its value.

- Moved the eval stuff out into dedicated functions so that they
are reusable and allow mutual recursion

- Some other cleanup relating to indentation