All functions are now variadic
All functions are now variadic. Functions accept a minimum number of arguments but
there is no maximum, extra arguments are implicitly pushed to the stack. This
simplifies the implementation greatly and also removes the division between
variadic and normal functions. A warning has been added for cases where you might
accidentally pass to many arguments to a function call. This change is mostly
backwards compatible except for variadic function definitions and variadic drop.

- Remove normal functions, all functions are variadic
- Add extra args warning