Scoped parameters and fix for segfault on string slicing
- Parameters are now scoped
This fixes an issue where you could modify a parameter through a nested call and the change would be persistant for the caller.

- String slicing segfault fix
The `begin` and `end` pointers would change due to the underlying std::string buffer being reallocated and would cause
these pointers to be invalidated thus causing a segfault due to accessing freed memory.

- Renamed `ctx` to `new`
- Removed use of `std::lower_bound` being used outside of function calls
Using `std::lower_bound` for things like drop and function definition didn't make sense because we wanted the exact function
in these cases and not the next best candidate.

 # Please enter the commit message for your changes. Lines starting