Error reporting overhaul
Big rewrite of error reporting system to make it a bit cleaner and to fix an issue
where errors occurring inside an eval context would use the line/column offset of
the string itself and not the location of the eval in its parent file.

Fixed an issue with error reporting segfaulting due to `wpp::Env` going out of scope in main.cpp

Print warnings only once.