Simplifying the lexer
A few changes to the lexer to simplify things and allow for greater flexibility in identifiers.
The lexer is now maximal munch and will always try to match the longest sequence of characters (usually).
For example, take the string `foo..bar`. Before, this would have concatenated two variables named `foo` and `bar`
but now it is actually a single identifier.

Other various changes have been made but wont have a visible effect for the end user but greatly simplify
the implementation.

Single line comments have been removed in favour of always using multiline comments.