Warning for deeply nested expressions
Add a new warning for deeply nested expressions in the parser.
An edge case that showed up when fuzzing with AFL is that deeply nested
expressions result in a stack overflow due to the use of a recursive
descent parser. While we can't really prevent this from happening, we
can at least report an error which may point to the cause being deeply
nested expressions.

A problem introduced by this warning was that we would end up seeing the
same warning printed repeatedly until the stack overflowed. The solution is
to use a tree structure to check if any parent nodes have already reported
the warning and if so, don't report it again.

To illustrate the solution better:

      1
     / \
    /   \
   2    [3] <- (a) warning about recursion happens here
  /     / \
 /     /   \
4     5     6
     /     / \
    /     /   \
   7     8    [9] <- (b) warning about recursion happens here

- Potential warning to be reported when we parse node 9

- Walk up tree to parent node 6

- Walk up tree to parent of 6 which is 3

- We have previously reported the same warning
  for node 3 which is an ancestor of node 9

- Don't show the warning for node 9

Summary of changes:
- New warning in the parser for deeply nested expressions
- Don't show errors that have already been reported for current or any parent nodes
- Introduction of a new data structure to keep track of child-to-parent relationships of nodes
- Add constants header to reduce code duplication